pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "3.7.7"
django = "^3.0.6"
django-extensions = "^2.2.9"
django-debug-toolbar = "^2.2"
werkzeug = "^1.0.1"
django-environ = "^0.4.5"
pillow = "^7.2.0"
django-imagekit = "^4.0.2"
django-registration-redux = "^2.8"
django-heroku = "^0.3.1"
gunicorn = "^20.0.4"
djangorestframework = "^3.11.0"
markdown = "^3.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
